Refactoring style folder
