+++ src/components/modals/NotFoundModal.js
@@ -1,7 +1,8 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import NotFoundPage from "../../pages/NotFoundPage";
-import { ModalBg, ModalBody, RelaitiveDiv } from "../../styled";
+import { ModalBg, ModalBody} from "../../style/modal";
+import { RelaitiveDiv } from "../../style/styled";
 
 const NotFoundModal = () => {
   const navigate = useNavigate();
