+++ src/components/modals/SignUpModal.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { ModalBg, ModalBody, RelaitiveDiv } from "../../styled";
+import { ModalBg, ModalBody} from "../../style/modal";
+import { RelaitiveDiv } from "../../style/styled";
 import SignUpForm from "../SignUpForm";
 
 const SignUpModal = () => {
