+++ src/pages/StartingPage.js
@@ -1,15 +1,13 @@
 import React from "react";
-import styled from "styled-components";
 import { Routes, Route } from "react-router-dom";
-import { devices } from "../device";
-import { ColumnFlexDiv } from "../styled";
+import { PageSection } from "../style/styled";
 import LogInModal from "../components/modals/LogInModal";
 import SignUpModal from "../components/modals/SignUpModal";
 import NotFoundModal from "../components/modals/NotFoundModal";
 
 const StartingPage = () => {
   return (
-    <StartingSection>
+    <PageSection>
       여기ㅏ는 스타팅 페이지입니다.
       <Routes>
         <Route index element={null}/>
@@ -17,16 +15,8 @@ const StartingPage = () => {
         <Route path="signup" element={<SignUpModal />} />
         <Route path="*" element={<NotFoundModal />} />
       </Routes>
-    </StartingSection>
+    </PageSection>
   );
 };
-const StartingSection = styled(ColumnFlexDiv)`
-  width: 100%;
-  height: 100%;
-  background-color:#FFF7C7;
-  @media ${devices.tablet} {
-    width: 100%;
-  }
-`;
 
 export default StartingPage;
